+++ docs/prelude.lua
@@ -48,41 +48,83 @@ Array = {
 }
 Array.__index = Array
 
+-- Constructs a new Array.
+-- @return [Array]
 function Array:new(init)
     local arr = {}
-    if (type(init) == 'table') then
+    if type(init) == 'table' then
         for _, v in ipairs(init) do
             table.insert(arr, v)
         end
     end
     setmetatable(arr, self)
+
     return arr
+end
+
+-- Constructs a new Array from either an other one or a table.
+-- @param other [Array|table]
+-- @return [Array]
+function Array:from(other)
+    if type(other) == 'table' then
+        return Array:new(other)
+    end
+
+    error('wrong format')
 end
 
 function Array:iterator()
     return ipairs(self)
 end
 
+-- Removes the last value of the Array represented by this one.
+-- @return [any]
 function Array:pop()
     local last_val = self[#self]
     table.remove(self)
+
     return last_val
 end
 
+-- Adds a new value at the end of the Array represented by this one.
+-- @param value [any]
+-- @return [any]
 function Array:push(value)
     table.insert(self, value)
+
     return value
 end
 
+-- Removes the first value of the Array represented by this one.
+-- @param value [any]
+-- @return [any]
 function Array:shift()
     local first_val = self[1]
     table.remove(self, 1)
+
     return first_val
 end
 
+-- Adds a new value at the front of the Array represented by this one.
+-- @param value [any]
+-- @return [any]
 function Array:unshift(value)
     table.insert(self, 1, value)
+
     return value
+end
+
+-- Creates a new Array populated with the results of calling a provided function
+-- on every element in the Array represented by this one.
+-- @param callback [function]
+-- @return [Array]
+function Array:map(callback)
+    local res = Array:new()
+    for _, v in ipairs(self) do
+        res:push(callback(v))
+    end
+
+    return res
 end
 
 -- Represents an Object with key insertion order iterator.
@@ -95,39 +137,62 @@ Object.__index = Object
 -- @return [Object]
 function Object:new(init)
     local obj = {
-        keys = {},   -- Array of keys (for insertion order).
-        values = {}, -- Map of key-value pairs.
+        _keys = {},     -- Array of keys (for insertion order).
+        _values = {},   -- Map of key-value pairs.
     }
-    if (type(init) == 'table') then
+    if type(init) == 'table' then
         for _, tuple in ipairs(init) do
             local k = tuple[1]
-            if (type(k) ~= 'string' or k:len() <= 0) then
+            if type(k) ~= 'string' or k:len() <= 0 then
                 error('wrong key format')
             end
 
-            if obj.values[k] == nil then
-                table.insert(obj.keys, k)
+            if obj._values[k] == nil then
+                table.insert(obj._keys, k)
             end
-            obj.values[k] = tuple[2]
+            obj._values[k] = tuple[2]
         end
     end
     setmetatable(obj, self)
+
     return obj
+end
+
+-- Constructs a new Object from either an other one or a table.
+-- @param other [Object|table]
+-- @return [Object]
+function Object:from(other)
+    local t = typeof(other)
+    if t == 'object' then
+        local new_obj = Object:new()
+        for _, e in other:iterator() do
+            new_obj:set(e.key, e.value)
+        end
+        return new_obj
+    elseif t == 'table' then
+        local new_obj = Object:new()
+        for k, v in pairs(other) do
+            new_obj:set(k, v)
+        end
+        return new_obj
+    end
+
+    error('wrong format')
 end
 
 -- Deletes a given key and its value.
 -- @param key [string]
 -- @return [any] The previous value if any, nil otherwise.
 function Object:delete(key)
-    for i, k in ipairs(self.keys) do
+    for i, k in ipairs(self._keys) do
         if k == key then
-            table.remove(self.keys, i)
+            table.remove(self._keys, i)
             break
         end
     end
 
-    local prev_val = self.values[key]
-    self.values[key] = nil
+    local prev_val = self._values[key]
+    self._values[key] = nil
 
     return prev_val
 end
@@ -136,24 +201,24 @@ end
 -- @param key [string]
 -- @return [any]
 function Object:get(key)
-    return self.values[key]
+    return self._values[key]
 end
 
 -- Tests whether a given key exists.
 -- @param key [string]
 -- @return [boolean]
 function Object:has(key)
-    return self.values[key] ~= nil
+    return self._values[key] ~= nil
 end
 
 function _object_iterator(obj, idx)
     local next_idx = idx + 1
-    local key = obj.keys[next_idx]
+    local key = obj._keys[next_idx]
     if key == nil then
         return
     end
 
-    local val = obj.values[key]
+    local val = obj._values[key]
     local entry = {
         key = key,
         value = val,
@@ -162,6 +227,7 @@ function _object_iterator(obj, idx)
     return next_idx, entry
 end
 
+-- Returns an entry iterator.
 function Object:iterator()
     return _object_iterator, self, 0
 end
@@ -172,7 +238,7 @@ end
 -- @param value [any]
 -- @return [any] The previous value if any, nil otherwise.
 function Object:set(key, value)
-    if (type(key) ~= 'string' or key:len() <= 0) then
+    if type(key) ~= 'string' or key:len() <= 0 then
         error('wrong key format')
     end
 
@@ -180,13 +246,13 @@ function Object:set(key, value)
         return self:delete(key)
     end
 
-    local prev_val = self.values[key]
+    local prev_val = self._values[key]
     if prev_val == nil then
         -- The given key does not exist yet.
-        table.insert(self.keys, key)
+        table.insert(self._keys, key)
     end
 
-    self.values[key] = value
+    self._values[key] = value
 
     return prev_val
 end
@@ -202,6 +268,29 @@ function Object:merge(other)
     return self
 end
 
+-- Returns an Array containing all keys in insertion order.
+-- @return [Array]
+function Object:keys()
+    local keys = Array:new()
+    for _, k in ipairs(self._keys) do
+        keys:push(k)
+    end
+
+    return keys
+end
+
+-- Returns an Array containing all values in insertion order.
+-- @return [Array]
+function Object:values()
+    local vals = Array:new()
+    for _, k in ipairs(self._keys) do
+        vals:push(self._values[k])
+    end
+
+    return vals
+end
+
+-- Represents a Context for accessing inputs and outputs.
 Context = {
     _classname = "Context",
 }