+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [Unreleased]
 ### Changed
 - Changed docs.
+### Added
+- Added Lua prelude functions:
+  - Array:from
+  - Array:map
+  - Object:from
+  - Object:keys
+  - Object:values
 
 ## [0.1.0] - 2022-01-09
 ### Added